pyproject.toml
[tool.poetry]
name = "underpeel-bot"
version = "0.1.0"
description = ""
authors = ["chezbgone <chezbgone@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.4.0"
aiohttp = "^3.11.11"
pydantic = "^2.10.6"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
sqlalchemy = "^2.0.38"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.3"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
